Rename folder and changes to emulation creator
Add docstrings
Refactor to use Command and Command list classes
Refactor to inherit from CommandCreator interface
